Rewrite of the activemq cookbook to resource based cookbook.
Based on the work done in tomcat, this provides an activemq_install and activemq_service resource.

An example of using these two resources can be seen in the test/cookbooks/test directory of this cookbook. It will set up an activemq_docs instance.

The *activemq_install* resource installs an instance of the activemq binary direct from Apache's mirror site. It is configured in such a way that if you wanted to run multiple broker instances on a single system you could, provided you explicitly set the ports to not conflict.

There are a multitude of configuration options that are available for ActiveMQ. This is just an initial install. The activemq.xml.erb template can be overridden to allow for security and other options.

Integration tests have been converted to inspec. An additional check has been added to test that the queue is working using the _activemq producer_ and _activemq consumer_ commands.

The community engineering team welcomes contributions to community cookbooks. We centralize our documentation for [contributing](https://github.com/chef-cookbooks/community_cookbook_documentation/blob/master/CONTRIBUTING.MD) which walks through the process.